+++ packages/components/src/styles/link-button-pure-styles.ts
@@ -1,12 +1,19 @@
 import type { Styles, JssStyle } from 'jss';
-import type { BreakpointCustomizable, GetStylesFunction } from '../utils';
+import type { BreakpointCustomizable, GetJssStyleFunction } from '../utils';
 import type { AlignLabel, AlignLabelType, LinkButtonPureIconName, TextSize, ThemeExtendedElectricDark } from '../types';
 import { buildResponsiveStyles, generateTypeScale, hasVisibleIcon, mergeDeep, paramCaseToCamelCase } from '../utils';
-import { addImportantToRule, getFocusStyles, getInset, getTransition, pxToRemWithUnit, getThemedColors } from './';
+import {
+  addImportantToRule,
+  getFocusJssStyle,
+  getInsetJssStyle,
+  getTransition,
+  pxToRemWithUnit,
+  getThemedColors,
+} from './';
 import { fontSize, getScreenReaderOnlyJssStyle } from '@porsche-design-system/utilities-v2';
 import { isSizeInherit } from '../components/basic/typography/text/text-utils';
 
-const getSizeStyles: GetStylesFunction = (textSize: TextSize): JssStyle => {
+const getSizeJssStyle: GetJssStyleFunction = (textSize: TextSize): JssStyle => {
   if (isSizeInherit(textSize)) {
     return {
       fontSize: 'inherit',
@@ -48,7 +55,7 @@ const getSizeStyles: GetStylesFunction = (textSize: TextSize): JssStyle => {
   }
 };
 
-const getVisibilityStyles: GetStylesFunction = (hideLabel: boolean): JssStyle => {
+const getVisibilityJssStyle: GetJssStyleFunction = (hideLabel: boolean): JssStyle => {
   return hideLabel
     ? getScreenReaderOnlyJssStyle()
     : {
@@ -64,7 +71,7 @@ const getVisibilityStyles: GetStylesFunction = (hideLabel: boolean): JssStyle =>
       };
 };
 
-const getLabelAlignmentStyles: GetStylesFunction = (alignLabel: AlignLabelType): JssStyle => {
+const getLabelAlignmentJssStyle: GetJssStyleFunction = (alignLabel: AlignLabelType): JssStyle => {
   return alignLabel === 'left'
     ? {
         padding: `0 ${pxToRemWithUnit(4)} 0 0`,
@@ -77,17 +84,17 @@ const getLabelAlignmentStyles: GetStylesFunction = (alignLabel: AlignLabelType):
 };
 
 /* Needed for slotted anchor and hidden label, which then enlarges the hidden label to equal host size and indents the text to be visually hidden. */
-const getSlottedAnchorVisibilityStyles: GetStylesFunction = (hideLabel: boolean): JssStyle => {
+const getSlottedAnchorVisibilityJssStyle: GetJssStyleFunction = (hideLabel: boolean): JssStyle => {
   return hideLabel
     ? {
         position: 'absolute',
-        ...getInset(),
+        ...getInsetJssStyle(),
         whiteSpace: 'nowrap',
         textIndent: '-999999px',
       }
     : {
         position: 'static',
-        ...getInset('auto'),
+        ...getInsetJssStyle('auto'),
         whiteSpace: 'inherit',
         textIndent: 0,
       };
@@ -138,7 +145,7 @@ export const getLinkButtonPureStyles = (
       background: 'transparent',
       color: isDisabledOrLoading ? disabledColor : active ? activeColor : baseColor,
       transition: `${getTransition('color')}, font-size 1ms linear`, // used for transitionend event listener
-      ...(!hasSlottedAnchor && getFocusStyles({ offset: 1, pseudo: '::before' })),
+      ...(!hasSlottedAnchor && getFocusJssStyle({ offset: 1, pseudo: '::before' })),
       ...(!isDisabledOrLoading && {
         '&:hover': {
           color: hoverColor,
@@ -162,7 +169,7 @@ export const getLinkButtonPureStyles = (
           buildResponsiveStyles(stretch, (stretched: boolean) => ({
             justifyContent: stretched ? 'space-between' : 'flex-start',
           })),
-        buildResponsiveStyles(size, getSizeStyles)
+        buildResponsiveStyles(size, getSizeJssStyle)
       ),
     },
     ...(hasIcon && {
@@ -173,8 +180,13 @@ export const getLinkButtonPureStyles = (
       },
       label: {
         ...mergeDeep(
-          buildResponsiveStyles(hideLabel, !hasSlottedAnchor ? getVisibilityStyles : getSlottedAnchorVisibilityStyles),
-          hasSubline ? { paddingLeft: pxToRemWithUnit(4) } : buildResponsiveStyles(alignLabel, getLabelAlignmentStyles)
+          buildResponsiveStyles(
+            hideLabel,
+            !hasSlottedAnchor ? getVisibilityJssStyle : getSlottedAnchorVisibilityJssStyle
+          ),
+          hasSubline
+            ? { paddingLeft: pxToRemWithUnit(4) }
+            : buildResponsiveStyles(alignLabel, getLabelAlignmentJssStyle)
         ),
       },
     }),
@@ -182,10 +194,10 @@ export const getLinkButtonPureStyles = (
       subline: {
         display: 'flex',
         transition: getTransition('color'),
-        marginTop: addImportantToRule('4px'), // override due to reset of getScreenReaderOnlyJssStyle() in getVisibilityStyles
+        marginTop: addImportantToRule('4px'), // override due to reset of getScreenReaderOnlyJssStyle() in getVisibilityJssStyle
         color: isDisabledOrLoading ? disabledColor : active ? activeColor : baseColor,
         ...(hasIcon && {
-          ...buildResponsiveStyles(hideLabel, getVisibilityStyles),
+          ...buildResponsiveStyles(hideLabel, getVisibilityJssStyle),
           paddingLeft: pxToRemWithUnit(4),
           '&::before': {
             content: '""',